Desenvolvimento do código para acesso ao banco de dados.

+++ REST-API - 01 Calculator/REST-API/Services/IPersonService.cs
@@ -9,6 +9,6 @@ namespace RESTAPI.Services
         Person FindById(long id);
         List<Person> FindAll();
         Person Update(Person person);
-        void Delete(long id);
+        void Delete(int id);
     }
 }
